Adding sealed keyword to improve code and its efficiency

+++ MightyMats/Areas/Admin/Controllers/ProductController.cs
@@ -6,11 +6,11 @@ using NToastNotify;
 namespace MightyMats.Controllers;
 
 [Area("Admin")]
-public class ProductController : Controller
+public sealed class ProductController : Controller
 {
-    private readonly IUnitOfWork _unitOfWork;
-    private readonly IToastNotification _toastNotification;
     private readonly IWebHostEnvironment _hostEnvironment;
+    private readonly IToastNotification _toastNotification;
+    private readonly IUnitOfWork _unitOfWork;
 
     public ProductController(IUnitOfWork unitOfWork, IToastNotification toastNotification,
         IWebHostEnvironment hostEnvironment)
@@ -23,7 +23,7 @@ public class ProductController : Controller
 
     public async Task<IActionResult> Index()
     {
-        IEnumerable<Product> productsList = await _unitOfWork._productRepository.GetAll(null);
+        var productsList = await _unitOfWork._productRepository.GetAll();
         return View(productsList);
     }
 
@@ -35,11 +35,11 @@ public class ProductController : Controller
     [HttpPost]
     public async Task<IActionResult> Create(Product product, IFormFile? file)
     {
-        string webRootPath = _hostEnvironment.WebRootPath;
+        var webRootPath = _hostEnvironment.WebRootPath;
         if (file != null)
         {
-            string fileName = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName);
-            string productsPath = Path.Combine(webRootPath, @"images/products");
+            var fileName = Guid.NewGuid() + Path.GetExtension(file.FileName);
+            var productsPath = Path.Combine(webRootPath, @"images/products");
 
             using (var fileStream = new FileStream(Path.Combine(productsPath, fileName), FileMode.Create))
             {
@@ -57,39 +57,33 @@ public class ProductController : Controller
             _toastNotification.AddSuccessToastMessage("Successful creation");
             return RedirectToAction(nameof(Index));
         }
-        else
-        {
-            return View(product);
-        }
+
+        return View(product);
     }
 
 
     public async Task<IActionResult> Edit(int? id)
     {
-        Product product = await _unitOfWork._productRepository.Get(_ => _.Id == id);
+        var product = await _unitOfWork._productRepository.Get(_ => _.Id == id);
         return View(product);
     }
 
     [HttpPost]
     public async Task<IActionResult> Edit(Product product, IFormFile? file)
     {
-
-        string webRootPath = _hostEnvironment.WebRootPath;
+        var webRootPath = _hostEnvironment.WebRootPath;
         if (ModelState.IsValid)
         {
             if (file != null)
             {
-                string fileName = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName);
-                string productsPath = Path.Combine(webRootPath, @"images/products");
+                var fileName = Guid.NewGuid() + Path.GetExtension(file.FileName);
+                var productsPath = Path.Combine(webRootPath, @"images/products");
 
                 if (!string.IsNullOrEmpty(product.Image))
                 {
-                    string oldImagePath = $"{webRootPath}{product.Image.TrimStart('\\')}";
+                    var oldImagePath = $"{webRootPath}{product.Image.TrimStart('\\')}";
 
-                    if (System.IO.File.Exists(oldImagePath))
-                    {
-                        System.IO.File.Delete(oldImagePath);
-                    }
+                    if (System.IO.File.Exists(oldImagePath)) System.IO.File.Delete(oldImagePath);
                 }
 
                 using (var fileStream = new FileStream(Path.Combine(productsPath, fileName), FileMode.Create))
@@ -106,26 +100,23 @@ public class ProductController : Controller
             _toastNotification.AddInfoToastMessage("Successful update");
             return RedirectToAction(nameof(Index));
         }
-        
+
         return NotFound();
     }
 
     public async Task<IActionResult> Delete(int? id)
     {
-        string webRootPath = _hostEnvironment.WebRootPath;
-        Product? product = await _unitOfWork._productRepository.Get(_ => _.Id == id);
+        var webRootPath = _hostEnvironment.WebRootPath;
+        var product = await _unitOfWork._productRepository.Get(_ => _.Id == id);
 
         if (product == null || product.Id == null || product.Id == 0)
             return NotFound();
 
         if (!string.IsNullOrEmpty(product.Image))
         {
-            string oldImagePath = $"{webRootPath}{product.Image.TrimStart('\\')}";
+            var oldImagePath = $"{webRootPath}{product.Image.TrimStart('\\')}";
 
-            if (System.IO.File.Exists(oldImagePath))
-            {
-                System.IO.File.Delete(oldImagePath);
-            }
+            if (System.IO.File.Exists(oldImagePath)) System.IO.File.Delete(oldImagePath);
         }
 
         if (ModelState.IsValid)
